finished lab 3

+++ LabIV/arrayutils.h
@@ -9,7 +9,7 @@ double dcheckit(char*);
 void cinsert(char*);
 void carelins(char*, int);
 void darelins(double*, int);
-int genrandom(int);
+double genrandom(int);
 void dfillarran(double*, int);
 void printa(double*, int);
 void printmx(double**, int, int);
